Move MultiDialog behind the BasicUI facade...
... So that it breaks dependency on ShuttleGui and the help system.

Perhaps this is stretching the BasicUI toolkit with a rather arbitrary and
complicated dialog.  Perhaps a better future toolkit neutral ShuttleGui and
help system in other libraries could make this go away.

Note that there is no other use of MultiDialog, now that there aren't orphan
block files.

The type ManualPageId is used to avoid any mention of "wx" in this interface.

===============================================================================

Cherry-picked from Audacity commit 3e41852. Modified for Tenacity.

Signed-off-by: Avery King <avery98@pm.me>